Refactor payvision-be to run correctly
