+++ .gitignore
@@ -1,4 +1,5 @@
 .env
 payvision.mdj
 package-lock.json
-node-modules/
+node_modules/
+conf_db.sql
