+++ payvision-be/server.js
@@ -2,7 +2,7 @@ require('dotenv').config();
 const express = require('express');
 
 const cors = require('cors');
-const db = require('./models');
+const {db, sequelize} = require('./models');
 
 const app = express();
 const port = process.env.PORT || 8000;
@@ -16,4 +16,9 @@ app.use('/api/auth', require('./routes/auth.route'));
 
 app.get('/', (req, res) => {res.send("Welcome to payvision backend!")});
 
-app.listen(port, () => {console.log(`server running port ${port}`);});
+sequelize
+  .sync({ alter: true })
+  .then((result) => {
+    app.listen(port, () => {console.log(`server running port ${port}`);});
+  })
+  .catch((err) => console.log(err));
